+++ packages/block-serialization-spec-parser/parser.js
@@ -166,7 +166,7 @@
             /** <?php
             return array(
               'blockName'    => $blockName,
-              'attrs'        => isset( $attrs ) ? $attrs : array(),
+              'attrs'        => empty( $attrs ) ? peg_empty_attrs() : $attrs,
               'innerBlocks'  => array(),
               'innerHTML'    => '',
               'innerContent' => array(),
@@ -187,7 +187,7 @@
 
             return array(
               'blockName'    => $s['blockName'],
-              'attrs'        => $s['attrs'],
+              'attrs'        => empty( $s['attrs'] ) ? peg_empty_attrs() : $s['attrs'],
               'innerBlocks'  => $innerBlocks,
               'innerHTML'    => $innerHTML,
               'innerContent' => $innerContent,
@@ -1618,6 +1618,18 @@
     // The `maybeJSON` function is not needed in PHP because its return semantics
     // are the same as `json_decode`
 
+    if ( ! function_exists( 'peg_empty_attrs' ) ) {
+         function peg_empty_attrs() {
+             static $empty_attrs = null;
+
+             if ( null === $empty_attrs ) {
+                 $empty_attrs = json_decode( '{}', true );
+             }
+
+             return $empty_attrs;
+         }
+    }
+
     // array arguments are backwards because of PHP
     if ( ! function_exists( 'peg_process_inner_content' ) ) {
         function peg_process_inner_content( $array ) {
@@ -1646,7 +1658,7 @@
             if ( ! empty( $pre ) ) {
                 $blocks[] = array(
                     'blockName' => null,
-                    'attrs' => array(),
+                    'attrs' => peg_empty_attrs(),
                     'innerBlocks' => array(),
                     'innerHTML' => $pre,
                     'innerContent' => array( $pre ),
@@ -1661,7 +1673,7 @@
                 if ( ! empty( $html ) ) {
                     $blocks[] = array(
                         'blockName' => null,
-                        'attrs' => array(),
+                        'attrs' => peg_empty_attrs(),
                         'innerBlocks' => array(),
                         'innerHTML' => $html,
                         'innerContent' => array( $html ),
@@ -1672,7 +1684,7 @@
             if ( ! empty( $post ) ) {
                 $blocks[] = array(
                     'blockName' => null,
-                    'attrs' => array(),
+                    'attrs' => peg_empty_attrs(),
                     'innerBlocks' => array(),
                     'innerHTML' => $post,
                     'innerContent' => array( $post ),